Fix #610, add support for vxworks7
This adds support for vxworks7 with its minor differences from vxworks6.
