+++ src/os/vxworks/src/os-impl-dirs-globals.c
@@ -19,36 +19,37 @@
  */
 
 /**
- * \file     os-impl-tasks.h
+ * \file     os-impl-dirs.c
  * \ingroup  vxworks
  * \author   joseph.p.hickey@nasa.gov
  *
  */
 
-#ifndef INCLUDE_OS_IMPL_TASKS_H_
-#define INCLUDE_OS_IMPL_TASKS_H_
+/****************************************************************************************
+                                    INCLUDE FILES
+****************************************************************************************/
 
-#include <osconfig.h>
-#include <taskLib.h>
+#include "os-vxworks.h"
+#include "os-impl-dirs.h"
+#include "os-shared-dir.h"
 
-#if defined(VX_WIND_TCB_SIZE)
-/* vxworks >= 7.0 should provide this symbol via taskLib.h. WIND_TCB is an opaque type */
-typedef char OS_VxWorks_TCB_t[VX_WIND_TCB_SIZE];
-#else
-/* older vxworks expose the definition of VX_WIND_TCB_SIZE */
-typedef WIND_TCB OS_VxWorks_TCB_t;
-#endif /* !defined(VX_WIND_TCB_SIZE) */
 
-/*tasks */
-typedef struct
+/*
+ * The directory handle table.
+ */
+OS_impl_dir_internal_record_t OS_impl_dir_table[OS_MAX_NUM_OPEN_DIRS];
+
+
+/*----------------------------------------------------------------
+ *
+ * Function: OS_VxWorks_DirAPI_Impl_Init
+ *
+ *  Purpose: Local helper routine, not part of OSAL API.
+ *
+ *-----------------------------------------------------------------*/
+int32 OS_VxWorks_DirAPI_Impl_Init(void)
 {
-    OS_VxWorks_TCB_t tcb; /* Must be first */
-    TASK_ID vxid;
-    void    *heap_block;    /* set non-null if the stack was obtained with malloc() */
-    size_t  heap_block_size;
-} OS_impl_task_internal_record_t;
+   memset(OS_impl_dir_table, 0, sizeof(OS_impl_dir_table));
+   return OS_SUCCESS;
+} /* end OS_VxWorks_DirAPI_Impl_Init */
 
-/* Tables where the OS object information is stored */
-extern OS_impl_task_internal_record_t OS_impl_task_table[OS_MAX_TASKS];
-
-#endif /* INCLUDE_OS_IMPL_TASKS_H_ */
