+++ src/unit-test-coverage/vxworks/src/coveragetest-dirs-globals.c
@@ -1,0 +1,76 @@
+/*
+ *  NASA Docket No. GSC-18,370-1, and identified as "Operating System Abstraction Layer"
+ *
+ *  Copyright (c) 2019 United States Government as represented by
+ *  the Administrator of the National Aeronautics and Space Administration.
+ *  All Rights Reserved.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
+/**
+ * \file     coveragetest-dirs-globals.c
+ * \ingroup  vxworks
+ * \author   steven.seeger@nasa.gov
+ *
+ */
+
+#include "os-vxworks-coveragetest.h"
+#include "ut-adaptor-dirs.h"
+
+#include "os-shared-dir.h"
+
+#include <OCS_stdlib.h>
+#include <OCS_taskLib.h>
+#include <OCS_dirent.h>
+#include <OCS_unistd.h>
+#include <OCS_stat.h>
+
+void Test_OS_VxWorks_DirAPI_Impl_Init(void)
+{
+    /*
+     * Test Case For:
+     * int32 OS_VxWorks_DirAPI_Impl_Init(void)
+     */
+    OSAPI_TEST_FUNCTION_RC(UT_Call_OS_VxWorks_DirAPI_Impl_Init(), OS_SUCCESS);
+}
+
+/* ------------------- End of test cases --------------------------------------*/
+
+/* Osapi_Test_Setup
+ *
+ * Purpose:
+ *   Called by the unit test tool to set up the app prior to each test
+ */
+void Osapi_Test_Setup(void)
+{
+    UT_ResetState(0);
+}
+
+/*
+ * Osapi_Test_Teardown
+ *
+ * Purpose:
+ *   Called by the unit test tool to tear down the app after each test
+ */
+void Osapi_Test_Teardown(void) {}
+
+/* UtTest_Setup
+ *
+ * Purpose:
+ *   Registers the test cases to execute with the unit test tool
+ */
+void UtTest_Setup(void)
+{
+    ADD_TEST(OS_VxWorks_DirAPI_Impl_Init);
+}
